implemented basic calendar view (no brushing & linking)
Implemented the basic calendar view that currently allows users to hover over the months and see which year they selected in the dropdown.
It is currently linked to the filter drop down menu in order to get the user selected year to show up in the header of the calendar.
(There isn't any brushing and linking done with the other visualizations in this implementation.)
@promiseadeliyi - please update the css for this as needed, thank you :D